feat: Files: Allows users to upload and share files.
fix: BREAKING CHANGE: runes to string converion now returns string, this mostly affect the validation responses. This makes the SDK compatible with Go 1.15

refactor: BREAKING CHANGE: EncryptString and DecryptString fuctions now accept a third param - bool, if true the IV is random and is sent along with the message. Default is false.